+++ forward_grpc_test.go
@@ -31,7 +31,7 @@ func newForwardGRPCFixture(t testing.TB, localConfig Config, sink sinks.MetricSi
 	// Create a global Veneur
 	globalCfg := globalConfig()
 	globalCfg.GrpcAddress = unusedLocalTCPAddress(t)
-	global := setupVeneurServer(t, globalCfg, nil, sink, nil)
+	global := setupVeneurServer(t, globalCfg, nil, sink, nil, nil)
 	go func() {
 		global.Serve()
 	}()
@@ -51,7 +51,7 @@ func newForwardGRPCFixture(t testing.TB, localConfig Config, sink sinks.MetricSi
 
 	localConfig.ForwardAddress = proxyCfg.GrpcAddress
 	localConfig.ForwardUseGrpc = true
-	local := setupVeneurServer(t, localConfig, nil, nil, nil)
+	local := setupVeneurServer(t, localConfig, nil, nil, nil, nil)
 
 	return &forwardGRPCFixture{t: t, proxy: &proxy, global: global, local: local}
 }
@@ -224,7 +224,8 @@ func TestE2EForwardingGRPCMetrics(t *testing.T) {
 		}
 
 		assert.ElementsMatch(t, expectedNames, actualNames,
-			"The global Veneur didn't flush the right metrics")
+			"The global Veneur didn't flush the right metrics: expectedNames=%v actualNames=%v",
+			expectedNames, actualNames)
 		close(done)
 	}()
 	ff.local.Flush(context.TODO())